requirements.txt
asgiref==3.5.2
Django==4.0.5
django-multiselectfield==0.1.12
Pillow==9.2.0
sqlparse==0.4.2
tzdata==2022.1
